chore: do change request fixes
