--- luxonis_ros/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/luxonis_ros
-[install]
-install_scripts=$base/lib/luxonis_ros
