+++ luxonis_scripts/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/luxonis_scripts
+[install]
+install_scripts=$base/lib/luxonis_scripts
