+++ rover-msgs/msg/ImageDetectionArray.msg
@@ -1,0 +1,1 @@
+ImageDetection[] detections
